Data.Monoid.First

Description

First' n is a generalization of the First exported by Data.Monoid: whereas Data.Monoid.First returns up to one value, First' n returns up to n values.

Data.Monoid.First    a ≡
            First' 1 a ≡
            First    a

This library also provides an API-compatible type synonym First and function getFirst allowing you to use it as a drop-in replacement.

First

type First a = First' 1 a

A type isomorphic to Data.Monoid.First

getFirst :: First a -> Maybe a

Get the first value of type a, if any.

>>> getFirst (foldMap pure [])
Nothing

>>> getFirst (foldMap pure [1,2,3,4])
Just 1

First'

data First' n a

A generalized version of Data.Monoid.First

getFirst' :: First' n a -> [a]

Get the first n values or fewer of type a.

>>> getFirst' (foldMap pure [1,2,3,4] :: First' 0 Int)
[]

>>> getFirst' (foldMap pure [1,2,3,4] :: First' 1 Int)
[1]

>>> getFirst' (foldMap pure [1,2,3,4] :: First' 2 Int)
[1,2]